refactor: Merge Apify RQ storage clients into one class
